annotation-system/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
tenacity>=8.0.0
tqdm>=4.62.0
argparse>=1.4.0